Updated the query to ensure the count is an integer, added documentation
The swagger file got more documentation to ensure that there are clear description fields for more of the components. This will be an ongoing process and one that doesn't really need any gitflow control.

+++ v2.0/helpers/landing/dbsummary.sql
@@ -6,7 +6,7 @@ select cdb.databasename,
   array_agg(distinct qt.datasettype) AS datasettypes,
   MIN(qt.younger) as younger,
   MAX(qt.older) as older,
-  COUNT(distinct qt.datasetid) AS datasets
+  COUNT(distinct qt.datasetid)::int AS datasets
 from ndb.constituentdatabases AS cdb
 left join ap.querytable as qt on cdb.databaseid = qt.databaseid
 left join ndb.contacts as cnt on cdb.contactid = cnt.contactid
